Extract magic __get method to ServiceNavigatorTrait
