+++ lib/Service/AbstractServiceFactory.php
@@ -14,11 +14,7 @@ namespace Stripe\Service;
  */
 abstract class AbstractServiceFactory
 {
-    /** @var \Stripe\StripeClientInterface */
-    private $client;
-
-    /** @var array<string, AbstractService|AbstractServiceFactory> */
-    private $services;
+    use ServiceNavigatorTrait;
 
     /**
      * @param \Stripe\StripeClientInterface $client
@@ -26,34 +22,5 @@ abstract class AbstractServiceFactory
     public function __construct($client)
     {
         $this->client = $client;
-        $this->services = [];
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return null|string
-     */
-    abstract protected function getServiceClass($name);
-
-    /**
-     * @param string $name
-     *
-     * @return null|AbstractService|AbstractServiceFactory
-     */
-    public function __get($name)
-    {
-        $serviceClass = $this->getServiceClass($name);
-        if (null !== $serviceClass) {
-            if (!\array_key_exists($name, $this->services)) {
-                $this->services[$name] = new $serviceClass($this->client);
-            }
-
-            return $this->services[$name];
-        }
-
-        \trigger_error('Undefined property: ' . static::class . '::$' . $name);
-
-        return null;
     }
 }
